oppdater mappestruktur for service
